requirements.txt
aiohttp==3.10.9
discord==2.3.2
discord.py==2.4.0
huggingface-hub==0.33.2
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.3
yt-dlp==2025.6.30
PyNaCl==1.5.0